+++ sites/geohub/src/components/VectorLayer.svelte
@@ -1,13 +1,12 @@
 <script lang="ts">
-  import { page } from '$app/stores'
   import { fade } from 'svelte/transition'
   import LayerNameGroup from '$components/control-groups/LayerNameGroup.svelte'
   import OpacityPanel from '$components/controls/OpacityPanel.svelte'
-  import VectorLegendPanel from '$components/controls/VectorLegendPanel.svelte'
+  import VectorLegend from '$components/controls/VectorLegend.svelte'
   import VectorLabelPanel from '$components/controls/VectorLabelPanel.svelte'
-  import { ClassificationMethodTypes, TabNames, VectorApplyToTypes } from '$lib/constants'
+  import { ClassificationMethodTypes, LegendTypes, TabNames, VectorApplyToTypes } from '$lib/constants'
   import type { Layer } from '$lib/types'
-  import VectorFilterPanelWizard from './controls/VectorFilterPanelWizard.svelte'
+  import VectorFilter from './controls/VectorFilter.svelte'
   import { Tabs } from '@undp-data/svelte-undp-design'
   import VectorParamsPanel from '$components/controls/VectorParamsPanel.svelte'
 
@@ -15,10 +14,8 @@
   export let classificationMethod: ClassificationMethodTypes
   export let colorMapName: string
 
-  const isReadonly = $page.url.pathname === '/viewer'
-
   let applyToOption: VectorApplyToTypes = VectorApplyToTypes.COLOR
-  let legendType: 'simple' | 'advanced'
+  let legendType: LegendTypes
   let defaultColor: string
   let defaultLineColor: string
   let activeTab = TabNames.LEGEND
@@ -30,11 +27,6 @@
     { label: TabNames.OPACITY, icon: 'fa-solid fa-droplet' },
     { label: TabNames.VECTORPARAMS, icon: 'fa-solid fa-sliders' },
   ]
-
-  if (isReadonly) {
-    tabs = [{ label: TabNames.OPACITY, icon: 'fa-solid fa-droplet' }]
-    activeTab = undefined
-  }
 </script>
 
 <div
@@ -54,7 +46,7 @@
 
     <p class="panel-content">
       {#if activeTab === TabNames.LEGEND}
-        <VectorLegendPanel
+        <VectorLegend
           {layer}
           bind:colorMapName
           bind:classificationMethod
@@ -63,7 +55,7 @@
           bind:defaultColor
           bind:defaultLineColor />
       {:else if activeTab === TabNames.FILTER}
-        <VectorFilterPanelWizard {layer} />
+        <VectorFilter {layer} />
       {:else if activeTab === TabNames.LABEL}
         <VectorLabelPanel {layer} />
       {:else if activeTab === TabNames.OPACITY}